cpp11armadillo: An 'Armadillo' Interface

Provides function declarations and inline function definitions that facilitate communication between R and the 'Armadillo' 'C++' library for linear algebra and scientific computing. This implementation is detailed in Vargas Sepulveda and Schneider Malamud (2024) doi:10.48550/arXiv.2408.11074 .

See also

Author

Maintainer: Mauricio Vargas Sepulveda m.sepulveda@mail.utoronto.ca (ORCID)

Authors:

Conrad Sanderson (Armadillo library (C++))

Other contributors:

Jonathan Schneider Malamud [contributor]